using vuex for serving gif
